Exports uglified version in dist
This is a workaround to the fact that browserify fails to recognize dependencies that have
already been browserified. Then, when this plugin is used elsewhere, there is a failure in which
the files local to qr-image are not found.

This problem with browserify has been reported and is being tracked on: substack/node-browserify#374
A temporary solution (this commit) is to use the uglified version, because the process changes
the "require" to some other crazy name and prevents the second browserify command to spoil the
references to the qr-image's files.